Update artifactory registry substitutor
