+++ dev/fattest.simplicity/resources/META-INF/services/org.testcontainers.dockerclient.DockerClientProviderStrategy
@@ -1,1 +1,1 @@
-componenttest.containers.ExternalTestServiceDockerClientStrategy
+componenttest.containers.ExternalDockerClientStrategy
