+++ dev/io.openliberty.org.testcontainers/cache/projects
@@ -146,6 +146,8 @@ com.ibm.ws.transaction.hadb_fat.sqlserver.retriablecodes
 fattest.simplicity
 io.openliberty.checkpoint_fat_container
 io.openliberty.checkpoint_fat_persistence
+io.openliberty.data.internal_fat
+io.openliberty.data.internal_fat_jpa
 io.openliberty.jcache.internal_fat.hazelcast
 io.openliberty.jcache.internal_fat.infinispan
 io.openliberty.microprofile.telemetry.1.0.internal.jaeger_fat
